+++ libnavui-maps/src/main/java/com/mapbox/navigation/ui/maps/route/line/model/MapboxRouteLineOptions.kt
@@ -6,8 +6,10 @@ import androidx.appcompat.content.res.AppCompatResources
 import com.mapbox.maps.extension.style.layers.properties.generated.IconAnchor
 import com.mapbox.maps.extension.style.layers.properties.generated.IconPitchAlignment
 import com.mapbox.maps.plugin.locationcomponent.LocationComponentConstants
+import com.mapbox.navigation.base.ExperimentalPreviewMapboxNavigationAPI
 import com.mapbox.navigation.ui.maps.route.RouteLayerConstants
 import com.mapbox.navigation.ui.maps.route.RouteLayerConstants.DEFAULT_ROUTE_SOURCES_TOLERANCE
+import com.mapbox.navigation.ui.maps.route.line.api.MapboxRouteLineView
 import com.mapbox.navigation.ui.maps.route.line.api.VanishingRouteLine
 import kotlin.math.abs
 
@@ -36,7 +38,12 @@ import kotlin.math.abs
  * @param waypointLayerIconOffset the list of offset values for waypoint icons
  * @param waypointLayerIconAnchor the anchor value, the default is [IconAnchor.CENTER]
  * @param iconPitchAlignment the pitch alignment value used for waypoint icons. The default is [IconPitchAlignment.MAP]
+ * @param shareLineGeometrySources enable route line's GeoJson source data sharing between multiple instances of the map.
+ * If this option is enabled for multiple instances of [MapboxRouteLineView]s that are used to draw route lines on multiple maps at the same time,
+ * they will all share the GeoJson source to optimize execution time of updates and decrease the memory footprint.
+ * **Enable only for instances that should share the geometry of the lines**, leave disabled for instances that should draw geometries distinct from other instances.
  */
+@OptIn(ExperimentalPreviewMapboxNavigationAPI::class)
 class MapboxRouteLineOptions private constructor(
     val resourceProvider: RouteLineResources,
     val routeStyleDescriptors: List<RouteStyleDescriptor>,
@@ -53,7 +60,9 @@ class MapboxRouteLineOptions private constructor(
         RouteLayerConstants.DEFAULT_VANISHING_POINT_MIN_UPDATE_INTERVAL_NANO,
     val waypointLayerIconOffset: List<Double> = listOf(0.0, 0.0),
     val waypointLayerIconAnchor: IconAnchor = IconAnchor.CENTER,
-    val iconPitchAlignment: IconPitchAlignment = IconPitchAlignment.MAP
+    val iconPitchAlignment: IconPitchAlignment = IconPitchAlignment.MAP,
+    @ExperimentalPreviewMapboxNavigationAPI
+    val shareLineGeometrySources: Boolean
 ) {
 
     /**
@@ -77,7 +86,8 @@ class MapboxRouteLineOptions private constructor(
             vanishingRouteLineUpdateIntervalNano,
             waypointLayerIconOffset,
             waypointLayerIconAnchor,
-            iconPitchAlignment
+            iconPitchAlignment,
+            shareLineGeometrySources
         )
     }
 
@@ -108,6 +118,7 @@ class MapboxRouteLineOptions private constructor(
         if (waypointLayerIconOffset != other.waypointLayerIconOffset) return false
         if (waypointLayerIconAnchor != other.waypointLayerIconAnchor) return false
         if (iconPitchAlignment != other.iconPitchAlignment) return false
+        if (shareLineGeometrySources != other.shareLineGeometrySources) return false
 
         return true
     }
@@ -130,6 +141,7 @@ class MapboxRouteLineOptions private constructor(
         result = 31 * result + (waypointLayerIconOffset.hashCode())
         result = 31 * result + (waypointLayerIconAnchor.hashCode())
         result = 31 * result + (iconPitchAlignment.hashCode())
+        result = 31 * result + (shareLineGeometrySources.hashCode())
         return result
     }
 
@@ -150,7 +162,8 @@ class MapboxRouteLineOptions private constructor(
             "vanishingRouteLineUpdateIntervalNano=$vanishingRouteLineUpdateIntervalNano," +
             "waypointLayerIconOffset=$waypointLayerIconOffset," +
             "waypointLayerIconAnchor=$waypointLayerIconAnchor," +
-            "iconPitchAlignment=$iconPitchAlignment" +
+            "iconPitchAlignment=$iconPitchAlignment," +
+            "shareLineGeometrySources=$shareLineGeometrySources" +
             ")"
     }
 
@@ -175,6 +188,10 @@ class MapboxRouteLineOptions private constructor(
      * @param vanishingRouteLineUpdateIntervalNano can be used to decrease the frequency of the vanishing route
      * line updates improving the performance at the expense of visual appearance of the vanishing point on the line during navigation.
      * @param iconPitchAlignment the pitch alignment value used for waypoint icons. The default is [IconPitchAlignment.MAP]
+     * @param shareLineGeometrySources enables route line's GeoJson source data sharing between multiple instances of the map.
+     * If this option is enabled for multiple instances of [MapboxRouteLineView]s that are used to draw route lines on multiple maps at the same time,
+     * they will all share the GeoJson source to optimize execution time of updates and decrease the memory footprint.
+     * **Enable only for instances that should share the geometry of the lines**, leave disabled for instances that should draw geometries distinct from other instances.
      */
     class Builder internal constructor(
         private val context: Context,
@@ -190,7 +207,8 @@ class MapboxRouteLineOptions private constructor(
         private var vanishingRouteLineUpdateIntervalNano: Long,
         private var iconOffset: List<Double>,
         private var iconAnchor: IconAnchor,
-        private var iconPitchAlignment: IconPitchAlignment
+        private var iconPitchAlignment: IconPitchAlignment,
+        private var shareLineGeometrySources: Boolean
     ) {
 
         /**
@@ -212,7 +230,8 @@ class MapboxRouteLineOptions private constructor(
             RouteLayerConstants.DEFAULT_VANISHING_POINT_MIN_UPDATE_INTERVAL_NANO,
             listOf(0.0, 0.0),
             IconAnchor.CENTER,
-            IconPitchAlignment.MAP
+            IconPitchAlignment.MAP,
+            false
         )
 
         /**
@@ -364,6 +383,19 @@ class MapboxRouteLineOptions private constructor(
             apply { this.iconPitchAlignment = iconPitchAlignment }
 
         /**
+         * Enable route line's GeoJson source data sharing between multiple instances of the map.
+         * If this option is enabled for multiple instances of [MapboxRouteLineView]s that are used to draw route lines on multiple maps at the same time,
+         * they will all share the GeoJson source to optimize execution time of updates and decrease the memory footprint.
+         * **Enable only for instances that should share the geometry of the lines**, leave disabled for instances that should draw geometries distinct from other instances.
+         *
+         * @param shareLineGeometrySources false by default
+         * @return the builder
+         */
+        @ExperimentalPreviewMapboxNavigationAPI
+        fun shareLineGeometrySources(shareLineGeometrySources: Boolean): Builder =
+            apply { this.shareLineGeometrySources = shareLineGeometrySources }
+
+        /**
          * @return an instance of [MapboxRouteLineOptions]
          */
         fun build(): MapboxRouteLineOptions {
@@ -401,7 +433,8 @@ class MapboxRouteLineOptions private constructor(
                 vanishingRouteLineUpdateIntervalNano,
                 iconOffset,
                 iconAnchor,
-                iconPitchAlignment
+                iconPitchAlignment,
+                shareLineGeometrySources
             )
         }
     }